Refactor: AdminPlaceValidator로 승인 검증 로직 이동
